Update patch set 1
Patch Set 1: Code-Review+1

Is there a way we can add more context to this request for donations? Having a link with no additional context does imply that the maintainers of the site you're looking at in some way endorse this org, but it leaves the reader wondering why this particular org is important to this project.

I wonder if we need a "project history" page or something on the gerritcodereview.com site, where we can talk about Shawn's involvement in the project and his legacy. In that context, donating to a cancer charity will make more sense.

Patch-set: 1
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1